Update SumaPolizas to decimal and add Excel export feature
- Changed SumaPolizas type from int? to decimal? in DeclaracionCfdiBase and updated related formatting in CFDITable.razor.
- Fixed parameter in LigarCFDIPolizaPage to use Comprobante.Total.
- Refactored ContabilidadRegistros.razor to add Excel export functionality using EPPlus, including UI changes and JavaScript interop for file download.
- Improved account picker event handling and UI layout for action buttons.
- Added helper methods for safe Excel file naming.

+++ src/OrionERP.Application/Features/Cfdi/DeclaracionPrevia/DeclaracionPreviaModels.cs
@@ -40,7 +40,7 @@ public class DeclaracionCfdiBase : IDeclaracionComprobanteItem
   public string? Estatus { get; set; }
   public string? fechastransacciones { get; set; }
   public string? Poliza { get; set; }
-  public int? SumaPolizas { get; set; }
+  public decimal? SumaPolizas { get; set; }
   public int? XML_Attachment_ID { get; set; }
   public bool EsEmitida { get; set; }
   public bool EsRecibida { get; set; }
